Adjust usage of raise and assert_raises
- replaces usage of raise and assert_raises from nose.tools to
  nose_utils
- updates:
  - dali/test/python/test_dali_tf_dataset_graph.py
  - dali/test/python/test_operator_arithmetic_ops.py
  - dali/test/python/test_operator_mfcc.py
  - dali/test/python/test_pool.py
  - dali/test/python/test_operator_roi_random_crop.py
  - dali/test/python/test_operator_expand_dims.py
  - dali/test/python/test_operator_slice.py
- unifies CPU and GPU asses in MFCC operator
- unifies unary and binary operator assert for arithmetic operators
- fixes test_tf_dataset_wrong_placement_cpu and test_tf_dataset_wrong_placement_gpu
  tests

+++ dali/operators/generic/expand_dims.cc
@@ -59,7 +59,7 @@ ExpandDims<Backend>::ExpandDims(const OpSpec &spec)
   if (!new_axis_names.empty()) {
     DALI_ENFORCE(new_axis_names.size() == axes.size(), make_string("Specified ", axes.size(),
       " new dimensions, but layout contains only ",
-      new_axis_names.size(), " new dimension names"));
+      static_cast<int>(new_axis_names.size()), " new dimension names"));
   }
   this->use_src_dims_ = true;
